refactor: renamed Custom Appbar from MUI
